Backend/requirements.txt
flask 
flask-cors 
supabase 
python-dotenv
gunicorn